Refactor checkbox to CourseCheckbox.tsx
